Rename files for factory-sensors
This exercise was previously named `errors`. For consistency, this
commit renames the solution and test file too.
