+++ exercises/concept/factory-sensors/factory-sensors.js
